Functional-ify day04 part02
So I figured a way out! This is much nicer than the original (frankly
pretty horrible) repeat approach, and just bumps counts based on
indexing. We don't need to keep all the info as we only care about the
ones coming after, so we can strip it off and bump counts like that.
Turns out to be somewhat neat!